pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "screener"
version = "0.1.0"
description = "Stock pattern screener for swing/position traders"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "yfinance>=0.2.30",
    "mplfinance>=0.12.9b7",
    "matplotlib>=3.7",
    "streamlit>=1.30",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "scipy>=1.11",
    "tqdm>=4.65",
    "scikit-learn>=1.3",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov"]

[tool.setuptools.packages.find]
where = ["src"]
